Introduce loader.py
